Ajuste no backend

+++ backend/src/main/java/sgc/diagnostico/model/AvaliacaoServidorRepo.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface AvaliacaoServidorRepo extends JpaRepository<AvaliacaoServidor, Integer> {
+public interface AvaliacaoServidorRepo extends JpaRepository<AvaliacaoServidor, Long> {
 
     @Query("""
         SELECT a FROM AvaliacaoServidor a
